Increased the buffer size for scanimage to 1 MB.
In other words, added --buffer-size=1024 to scanimage command line arguments. Increasing the buffer size was unnecessary when using scanimage from terminal; it successfully scans over 20 pages at a time. But when the command is executed in Python, with the default buffer size, the scanner stops working only after scanning a couple of pages.